Added a sensor
I added a sensor to make the robot move up when it encounters the floor, after that, the robot can move up and down automatically in specific height. Many things can be refined. I have two ideas. First, I want to add two sensors, one sensor is facing upwards, the other sensor is facing downwards. When any one encounters an obstacle, the wheels will reverse, but I think there may be no obstacles above the pole in reality, which makes it difficult to tell the robot "it's the time to move down", so I haven't made an upward sensor yet. My another idea is to use the height of the robot and the pole to tell the robot when to reverse the direction, but I checked the API document. We cannot seem to call the height of the robot in programming.
This version's problem: The detection distance of the downward sensor is too close so that the robot will hit the ground. I don't know how to adjust it.